Fix fleetshard central creation (#58)
