+++ dev/scripts/create-central.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 set -eo pipefail
 
+echo "Created central tenant: test-central-1"
+
 curl -X POST -H "Authorization: Bearer $(ocm token)" -H "Content-Type: application/json" \
   http://127.0.0.1:8000/api/rhacs/v1/centrals\?async\=true \
-  -d '{"name": "test-dinosaur-1", "multi_az": true, "cloud_provider": "standalone", "region": "standalone"}'
+  -d '{"name": "test-central-1", "multi_az": true, "cloud_provider": "standalone", "region": "standalone"}'
