+++ dev/scripts/ffm_curl.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+set -eo pipefail
+
+resource=${1}
+shift
+curl -v -H "Authorization: Bearer $(ocm token)" http://localhost:8000/api/${resource} $@
