+++ dev/scripts/setup-dev-env.sh
@@ -6,3 +6,6 @@ DIRNAME=$(dirname "$SCRIPT")
 
 echo "Copying cluster configuration"
 cp "$DIRNAME/../config/dataplane-cluster-configuration-minikube.yaml" "$DIRNAME/../../config/dataplane-cluster-configuration.yaml"
+
+# TODO: create Central request
+# ./"$DIRNAME"/create-central.sh
