+++ internal/dinosaur/pkg/handlers/data_plane_dinosaur.go
@@ -7,8 +7,8 @@ import (
 	"github.com/stackrox/acs-fleet-manager/pkg/handlers"
 	"net/http"
 
-	"github.com/stackrox/acs-fleet-manager/pkg/errors"
 	"github.com/gorilla/mux"
+	"github.com/stackrox/acs-fleet-manager/pkg/errors"
 )
 
 type dataPlaneDinosaurHandler struct {
