+++ scripts/local_db_setup.sh
